requirements.txt
openai
langchain
huggingface_hub
python_dotenv
streamlit